pyproject.toml
[project]
name = "transact-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "fastapi-users[sqlalchemy]>=14.0.0",
    "fastapi[standard]>=0.115.5",
    "geodantic>=0.4.0",
    "geojson-pydantic>=1.1.2",
    "geopandas>=1.0.1",
    "jinja2>=3.1.4",
    "owslib>=0.32.0",
    "playwright>=1.46.0",
    "psycopg[binary]>=3.2.3",
    "pydantic-settings>=2.6.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "tenacity>=9.0.0",
]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]
